requirements.txt
openai
pyyaml
tweepy
requests
bs4